requirements.txt
backports.tarfile==1.2.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
docutils==0.21.2
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.12.0
huggingface-hub==0.27.1
idna==3.10
importlib_metadata==8.5.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
Jinja2==3.1.5
keyring==25.6.0
kiwisolver==1.4.8
lib==4.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplot==0.1.9
matplotlib==3.10.0
mdurl==0.1.2
more-itertools==10.6.0
mpmath==1.3.0
networkx==3.4.2
nh3==0.2.20
numpy==2.2.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
pkginfo==1.12.0
pycparser==2.22
Pygments==2.19.1
pyloco==0.0.139
pyparsing==3.2.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
readme_renderer==44.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
safetensors==0.5.2
SecretStorage==3.3.3
SimpleWebSocketServer==0.1.2
six==1.17.0
sympy==1.13.1
tokenizers==0.21.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.48.1
triton==3.1.0
twine==6.0.1
typing==3.7.4.3
typing_extensions==4.12.2
urllib3==2.3.0
ushlex==0.99.1
websocket-client==1.8.0
zipp==3.21.0
